finwise_backend/requirements.txt
Django>=5.0

djangorestframework>=3.15

django-cors-headers>=4.3

djangorestframework-simplejwt>=5.3

# AI Service
google-generativeai>=0.8.5

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0 